Rewrite of the update-changes script.
We now use just 'git describe' to determine the current version
number. This generates versions of the form "<most-recent-tag>.<n>",
with n being the number of commits. While that's similar to the old
script, '<n>' is going to increase more often (or make larger jumps)
because git-describe counts commits on all paths leading to HEAD. But
it's much faster than the old script, and also more well-defined.

The new script can also call an arbitrary shell function with the new
version number, which allows to automatically update other files like
READMEs or scripts. Will add that to the repositories later as needed.

The script is still considered "experimental" and will probably
require a bit more tweaking.