Change utc options to not always override local time
Changed the utc options -u and -U so that they don't always override the
local time options -d and -D when both kinds of options are specified.
Instead, the last type of option (local or utc) specified takes effect (this
is consistent with what happens when both -c and -C options are given).

Also updated the usage message to make it more clear that the BRO_CUT_TIMEFMT
environment variable affects only the -u and -d options.