Fix bug in bro-cut when separator is not hexadecimal
Bro-cut was unable to process log files that had a literal character (instead
of a hexadecimal code) in the "#separator" line of the log file.  For
example, if a log file had a comma as the separator then bro-cut would
show either an error message or just blank lines, depending on the bro-cut
options being used.