Fix bug in bro-cut output of "#types" header line
Fixed a bug in bro-cut that could occur if both a time conversion option was
selected and a show header (-c or -C) option was selected.  In that
case, if the column number of the timestamp field in the log didn't match
the column number of the timestamp field on the command-line, then bro-cut
would not produce the correct output of the "#types" header line (it would
replace the wrong field with the string "string").  The fix for this bug
is to use the correct index value in the "times" array (the index must be
the column number in the log file).