Fix bug in bro-cut when log file has missing field
When bro-cut is processing multiple log files, then each time a log file
header is processed, bro-cut needs to clear the "columns" array, otherwise
there could be entries from a previous log file for fields that don't
exist in the current log file, which results in duplicate output columns
with no warning to the user.

This fix enables bro-cut to consistently output the "unknown field" error
message regardless of the order in which it processes such log files.