Fix bug in bro-cut when duplicate fields are specified
When duplicate field names were specified as bro-cut command-line parameters,
then bro-cut was producing bad output, because two identical column names
are the same index in the "idx" array, and so then the "columns" array
would have one undefined element.  Fixed by checking for this condition
and showing the user an error message.