Fix replace_version_in_rst function in update-changes script
When a reST document had a version string that contains the string
"beta", the update-changes script didn't recognize that string
as being part of the version string.  As a result, each time
update-changes was run it updated the version but would leave the "beta"
on the end, so the version string then contained multiple "beta" strings
which needed to be fixed manually.  See git commit 6a62d12b in the
broctl repo for an example.