Fix the handling of failed timestamp conversions
Fixed the case where an invalid timestamp was being output without
a separator character, and improved an existing test case to catch
this.

Fixed the case where the time format string was empty and bro-cut
complained that the timestamp was too long.  Now the error message
is correct.

When a formatted timestamp is too long, now bro-cut just outputs
the field without modification because strftime does not leave
the buffer in a well-defined state (i.e., trying to truncate it
doesn't always work).