Fix bro-cut to allow unset or zero time values
When doing time conversion, bro-cut would output an error message if it
encountered a time value of 0.0 or if the time value was not specified
(i.e., the unset field string).  An example is the kerberos.log which
has a time field with the &optional attribute.  Fixed so that bro-cut
no longer complains about these, and improved the time conversion
error messages.

Also added new tests and fixed the diff-time-zone canonifier script
to handle multiple timestamps on one line.