Add a time conversion test with duplicate timestamps
Changed the sample log file multiple-times.log to include cases where
the same time value is in multiple columns on one line, and the same
time value is in last time column of a line and the first time column
of the next line.

Added a new test case in time-conversion.test that uses the
log file multiple-times.log.