Faster bro-cut timestamp conversion
In Bro logs two consecutive timestamps often differ by less than a
second, so now bro-cut always stores the timestamp (and the
human-readable string) so that it can skip doing the timestamp
conversion when the next timestamp is the same.  This results in a
significant speedup when doing timestamp conversions on large Bro logs.