update-changes: expand the CHANGES initialization to simplify bootstrapping
This expands the existing init_changes function to look for an initial version
number in the VERSION file (which may already exist), and falls back to asking
the user to input one. This then triggers a first commit that installs the
CHANGES file and gets tagged with the provided version.