Do not explicitly add VERSION to `zeek_plugin_dist_files`.
`zeek-plugin-create-package.sh` already implicitly adds `VERSION` from
the plugin root to the distribution. Adding it explicitly caused us to
copy it to the build directory; since the build directory is always in
the C++ include lookup path this caused a conflict with C++ stdlib's
`<version>` header (in the C++ standard since only C++20, but for a
number of installations still searched for even when compiling with an
earlier standard).

Closes zeek/zeek#2784.