Add support for parsing log_suffix metadata from filenames
This is similar to the approach that we took for ZeekControl, tagging a
log_suffix to the archived file. With zeek-archiver, however, we can't
inject an environment variable that denotes the logger - there's only a
single zeek-archiver instance. Instead, the logger will encode that
information into filename:

    test__2020-07-16-09-43-10__2020-07-16-09-43-10__log_suffix=logger-1__.log

Outside of somehow adding a side-channel, the filename is all we have to
propagate this information. I didn't like the idea of having the position
imply the meaning, so key=value it became. It currently only recognizes
the log_suffix metadata.