Ignore merges of local branches in `update-changes`
We previously would ignore merges of remote branches when creating
CHANGELOG entries, but let merges of local branches through. This was due
to `update-changes` identifying merges through their Git-generated
commit message which assumes that the message was not edited.

With this patch we add proper detection of merge commits in that any
commit with more than a single parent is handled as a merge and excluded
from the generated `CHANGELOG`.