Replace gen-ct-list.pl with python version
The benefit here is that the python version uses entirely standard
library modules, which means a user won't need to install a bunch
of perl modules to get things going. Plus, we don't have a whole lot
of perl left in the ecosystem and this removes yet another instance
of it.